+++ src/main/java/ru/clevertec/service/card/CardServiceImpl.java
@@ -57,9 +57,9 @@ public class CardServiceImpl implements CardService {
     }
 
     private Result<String> getById(String idText) {
-        int id;
+        long id;
         try {
-            id = Integer.parseInt(idText);
+            id = Long.parseLong(idText);
         } catch (NumberFormatException e) {
             return new Fail<>(String.format(WRONG_ID_FORMAT_MESSAGE, idText), REQUEST_ERROR_CODE);
         }
@@ -138,11 +138,11 @@ public class CardServiceImpl implements CardService {
         else return new Fail<>(EMPTY_PRODUCT_LIST_MESSAGE, REQUEST_ERROR_CODE);
     }
 
-    private boolean isNumberNegativeOrZero(int number) {
+    private boolean isNumberNegativeOrZero(long number) {
         return number <= 0;
     }
 
-    private Result<String> removeCardById(int id) {
+    private Result<String> removeCardById(long id) {
         if (isNumberNegativeOrZero(id)) return new Fail<>(String.format(NEGATIVE_ID_MESSAGE, id), REQUEST_ERROR_CODE);
         boolean result;
         try {
